remove floating point

+++ benchmark/executor.py
@@ -101,7 +101,7 @@ def receive(exchange_name, routing_key):
             )
         )
         with open("{}".format(exchange_name), "w") as f:
-            f.write(str(LIMIT_UPDATES / (stop - start)))
+            f.write(str(int(LIMIT_UPDATES / (stop - start))))
 
     print("[+] Receiving {} on {}:{}".format(LIMIT_UPDATES, exchange_name, routing_key))
     with Connection(RABBITMQ_URI) as connection:
